renamed param_options.py to constants.py
