+++ nshm_logic_tree_utilities/scripts/plotting_script.py
@@ -2,7 +2,7 @@
 This script provides examples of how to use the plotting functions in plotting_functions.py.
 """
 
-import nshm_logic_tree_utilities.lib.param_options as param_options
+import nshm_logic_tree_utilities.lib.constants as param_options
 import nshm_logic_tree_utilities.lib.plotting_functions as plotting_functions
 
 plotting_functions.make_figure_of_coefficient_of_variation(
